Fixed bug related to log object existing after the first image
When only compressing images (not resizing), the code was looking to see if the log object exists, and if it did, it would get the name of the file from that object. This is correct for the case when we are also resizing, but in this case it is incorrect because it meant if we compress multiple images, subsequent compressions will always refer to the previous image.